[HMTM-187] - Naming, logic, factoring update
