+++ requirements.txt
@@ -1,5 +1,7 @@
+aiofiles==24.1.0
 aiohappyeyeballs==2.6.1
 aiohttp==3.11.16
+aiologger==0.7.0
 aiosignal==1.3.2
 annotated-types==0.7.0
 anyio==4.9.0
