+++ src/profile/constants.py
@@ -1,0 +1,50 @@
+from typing import Dict
+
+REQUEST_ENVIRONMENTS_MAPPING: Dict[str, str] = {
+    "user_not_found": "Пользователь не найден!",
+
+    "Пароль не соответствует требованиям: 8+ символов, A-Z, a-z, 0-9, спецсимвол.":
+        "Пароль не соответствует требованиям: 8+ символов, A-Z, a-z, 0-9, спецсимвол.",
+
+    "Имя пользователя не может быть короче 4-х символов":
+        "Имя пользователя не может быть короче 4-х символов!",
+
+    "Некорректное имя пользователя для telegram":
+        "Некорректное имя пользователя для telegram!",
+
+    "Некорректный номер телефона":
+        "Некорректный номер телефона!",
+
+    "Вы успешно поменяли данные о себе":
+        "Вы успешно поменяли данные о себе",
+
+    "Вы успешно поменяли данные о мастере":
+        "Вы успешно поменяли данные о мастере",
+
+    "Вы успешно стали мастером!":
+        "Вы успешно стали мастером!",
+
+    "Вы успешно поменяли пароль!":
+        "Вы успешно поменяли пароль!",
+
+    "Вы уже зарегистрированы!":
+        "Вы уже зарегистрированы!",
+
+    "У вас активная сессия!":
+        "У вас активная сессия!",
+
+    "Ваша почта уже подтверждена!":
+        "Ваша почта уже подтверждена!",
+
+    "Для смены забытого пароля вам необходима форма в профиле":
+        "Для смены забытого пароля вам необходима форма в профиле",
+
+    "Письмо о смене пароля отправлено на почту, указанную при регистрации!":
+        "Письмо о смене пароля отправлено на почту, указанную при регистрации!",
+
+    "Пользователь не найден":
+        "Пользователь не найден",
+
+    'Необходима почта, указанная при регистрации':
+        'Необходима почта, указанная при регистрации'
+}
