+++ src/profile/dto.py
@@ -1,0 +1,35 @@
+from dataclasses import dataclass
+from typing import Optional
+
+from src.dto import BoolResponse
+from src.profile.models import Master
+
+
+@dataclass
+class UpdateUserProfileResponse(BoolResponse):
+    pass
+
+
+@dataclass
+class ChangePasswordResponse(BoolResponse):
+    pass
+
+
+@dataclass
+class ChangeForgetPasswordResponse(BoolResponse):
+    pass
+
+
+@dataclass
+class GetUserIsMasterResponse(BoolResponse):
+    master: Optional[Master] = None
+
+
+@dataclass
+class UpdateMasterResponse(BoolResponse):
+    pass
+
+
+@dataclass
+class RegisterMasterResponse(BoolResponse):
+    pass
