+++ src/routes.py
@@ -7,7 +7,7 @@ from fastapi.templating import Jinja2Templates
 from src.cookies import set_cookie
 from src.sso.dto import GetMeResponse
 from src.sso import get_me as get_me_dependency
-from src.request_utils import extract_url_error_message
+from src.utils import extract_url_error_message
 
 main_router = APIRouter(tags=["Main"])
 templates = Jinja2Templates(directory="templates")
