+++ src/sso/routes.py
@@ -6,7 +6,6 @@ from fastapi.templating import Jinja2Templates
 
 from src.cookies import set_cookie
 from src.sso.dependencies import (
-    encryptor as encryptor_dependency,
     process_register as process_register_dependency,
     process_login as process_login_dependency,
     verify_email as verify_email_dependency,
@@ -14,14 +13,7 @@ from src.sso.dependencies import (
     send_verify_email_message as send_verify_email_message_dependency,
     send_forget_password_message as send_forget_password_message_dependency,
     change_forget_password as change_forget_password_dependency,
-    change_password as change_password_dependency,
-    update_user_profile as update_user_profile_dependency,
-    master_by_user as master_by_user_dependency,
-    update_master as update_master_info_dependency,
-    register_master as register_master_dependency,
     get_user_info as get_user_info_dependency,
-    toys_catalog as toys_catalog_dependency,
-    toy_by_id as toy_by_id_dependency,
 )
 from src.sso.dto import (
     LoginResponse,
@@ -30,21 +22,14 @@ from src.sso.dto import (
     VerifyEmailResponse,
     SendVerifyEmailMessageResponse,
     SendForgetPasswordMessageResponse,
-    ChangeForgetPasswordResponse,
-    ChangePasswordResponse,
-    UpdateUserProfileResponse,
-    GetUserIsMasterResponse,
-    UpdateMasterResponse,
-    RegisterMasterResponse,
     GetFullUserInfoResponse,
-    ToysCatalogResponse,
-    ToyByIDResponse,
 )
+from src.profile.dto import ChangeForgetPasswordResponse
 from src.sso.constants import FORGET_PASSWORD_TOKEN_NAME
-from src.request_utils import (
+from src.utils import (
     FernetEnvironmentsKey,
     extract_url_error_message,
-    extract_url_success_status_message
+    extract_url_success_status_message, encryptor as encryptor_dependency
 )
 
 router = APIRouter(prefix="/sso", tags=["SSO"])
@@ -156,38 +141,6 @@ async def verify_email(
     )
 
 
-@router.get("/profile", response_class=HTMLResponse, name="profile")
-async def profile_page(
-        request: Request,
-        current_user: GetMeResponse = Depends(get_me_dependency)
-):
-    if current_user.user is not None:
-        master: GetUserIsMasterResponse = await master_by_user_dependency(
-            user_id=current_user.user.id,
-            request=request,
-            cookies=current_user.cookies,
-        )
-
-        response: Response = templates.TemplateResponse(
-            request=request,
-            name="profile.html",
-            context={
-                "tab": request.query_params.get("tab") if request.query_params.get("tab") else "main",
-                "user": current_user.user,
-                "master": master.master if master.master else None,
-                "error_message": extract_url_error_message(request=request),
-                "success_message": extract_url_success_status_message(request=request)
-            }
-        )
-
-        for cookie in current_user.cookies:
-            response = set_cookie(response, cookie)
-
-        return response
-
-    return RedirectResponse(url="/sso/login", status_code=status.HTTP_303_SEE_OTHER)
-
-
 @router.get("/logout", response_class=RedirectResponse, name="logout")
 async def logout(request: Request):
     """ Завершение сессии """
@@ -276,12 +229,12 @@ async def process_send_forget_password_message(
         if result.error is not None:
             encrypted_error = encryptor.encrypt("Необходима почта, указанная при регистрации")
             return RedirectResponse(
-                url=f"/sso/profile?error={encrypted_error}&tab=security",
+                url=f"/profile/me?error={encrypted_error}&tab=security",
                 status_code=status.HTTP_303_SEE_OTHER
             )
 
         return RedirectResponse(
-            url=f"/sso/profile?message={encrypted_success_message}&tab=security",
+            url=f"/profile/me?message={encrypted_success_message}&tab=security",
             status_code=status.HTTP_303_SEE_OTHER
         )
 
@@ -342,132 +295,6 @@ async def process_change_forget_password(
     return response
 
 
-@router.post("/change-password", response_class=HTMLResponse, name="change-password")
-async def process_change_password(
-        current_user: GetMeResponse = Depends(get_me_dependency),
-        result: ChangePasswordResponse = Depends(change_password_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
-):
-    """ Ручка авторизованного пользователя для смены пароля - Процесс """
-    if current_user.user is None:
-        encrypted_error: str = encryptor.encrypt(str(current_user.error))
-        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
-
-    if result.error is not None:
-        encrypted_error_key: str = encryptor.encrypt(str(result.error))
-
-        return RedirectResponse(
-            url=f"/sso/profile?error={encrypted_error_key}&tab=security",
-            status_code=status.HTTP_303_SEE_OTHER
-        )
-
-    encrypted_message: str = encryptor.encrypt("Вы успешно поменяли пароль!")
-    response: Response = RedirectResponse(
-        url=f"/sso/profile?message={encrypted_message}&tab=security",
-        status_code=status.HTTP_303_SEE_OTHER
-    )
-
-    for cookie in result.cookies:
-        response = set_cookie(response, cookie)
-
-    return response
-
-
-@router.post("/update_user_profile", response_class=HTMLResponse, name="update_user_profile")
-async def process_update_user_profile(
-        request: Request,
-        result: UpdateUserProfileResponse = Depends(update_user_profile_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
-):
-    """ Ручка авторизованного пользователя для изменения данных о себе: никнейм, телеграм, телефон - Процесс """
-    current_user: GetMeResponse = await get_me_dependency(request=request, cookies=result.cookies)
-
-    if current_user.user is None:
-        encrypted_error: str = encryptor.encrypt(str(current_user.error))
-        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
-
-    if result.error is not None:
-        encrypted_error: str = encryptor.encrypt(str(result.error))  # type: ignore[no-redef]
-
-        return RedirectResponse(
-            url=f"/sso/profile?error={encrypted_error}&tab=main",
-            status_code=status.HTTP_303_SEE_OTHER
-        )
-
-    encrypted_message: str = encryptor.encrypt("Вы успешно поменяли данные о себе")
-    response: Response = RedirectResponse(
-        url=f"/sso/profile?message={encrypted_message}&tab=main",
-        status_code=status.HTTP_303_SEE_OTHER
-    )
-
-    for cookie in result.cookies:
-        response = set_cookie(response, cookie)
-
-    return response
-
-
-@router.post("/update-master", response_class=HTMLResponse, name="update-master")
-async def process_update_master(
-        current_user: GetMeResponse = Depends(get_me_dependency),
-        result: UpdateMasterResponse = Depends(update_master_info_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
-):
-    if current_user.user is None:
-        encrypted_error: str = encryptor.encrypt(str(current_user.error))
-        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
-
-    if result.error is not None:
-        encrypted_error: str = encryptor.encrypt(str(result.error))  # type: ignore[no-redef]
-
-        return RedirectResponse(
-            url=f"/sso/profile?error={encrypted_error}&tab=master",
-            status_code=status.HTTP_303_SEE_OTHER
-        )
-
-    encrypted_message: str = encryptor.encrypt("Вы успешно поменяли данные о мастере")
-    response: Response = RedirectResponse(
-        url=f"/sso/profile?message={encrypted_message}&tab=master",
-        status_code=status.HTTP_303_SEE_OTHER
-    )
-
-    for cookie in result.cookies:
-        response = set_cookie(response, cookie)
-
-    return response
-
-
-@router.post("/register-master", response_class=HTMLResponse, name="register-master")
-async def register_master(
-        current_user: GetMeResponse = Depends(get_me_dependency),
-        result: RegisterMasterResponse = Depends(register_master_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
-):
-    if current_user.user is None:
-        encrypted_error: str = encryptor.encrypt(str(current_user.error))
-        return RedirectResponse(f"/sso/login?error={encrypted_error}", status_code=status.HTTP_303_SEE_OTHER)
-
-    if result.error is not None:
-        encrypted_error: str = encryptor.encrypt(str(result.error))  # type: ignore[no-redef]
-
-        return RedirectResponse(
-            url=f"/sso/profile?error={encrypted_error}&tab=master",
-            status_code=status.HTTP_303_SEE_OTHER
-        )
-
-    encrypted_message: FernetEnvironmentsKey = FernetEnvironmentsKey()
-    encrypted_message_key: str = encrypted_message.encrypt("Вы успешно стали мастером!")
-
-    response: Response = RedirectResponse(
-        url=f"/sso/profile?message={encrypted_message_key}&tab=master",
-        status_code=status.HTTP_303_SEE_OTHER
-    )
-
-    for cookie in result.cookies:
-        response = set_cookie(response, cookie)
-
-    return response
-
-
 @router.get("/find-users", response_class=HTMLResponse, name="find_users")
 async def find_users_page(
         request: Request,
@@ -485,7 +312,7 @@ async def find_users_page(
 
 
 @router.post("/users/{query_params}", response_class=HTMLResponse, name="get_user_info")
-async def find_get_user_info(
+async def get_user_info(
         request: Request,
         result: GetFullUserInfoResponse = Depends(get_user_info_dependency),
         encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
@@ -504,67 +331,3 @@ async def find_get_user_info(
     }
 
     return templates.TemplateResponse(request=request, name="user-info.html", context=context)
-
-
-@router.get(path="/toys/catalog", response_class=HTMLResponse, name="toys_catalog")
-async def toys_catalog(
-        request: Request,
-        result: ToysCatalogResponse = Depends(toys_catalog_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
-):
-    current_user: GetMeResponse = await get_me_dependency(request=request)
-
-    if result.error is not None:
-        encrypted_error: str = encryptor.encrypt(result.error)
-
-        return RedirectResponse(
-            url=f"/?error={encrypted_error}",
-            status_code=status.HTTP_303_SEE_OTHER
-        )
-
-    context = {
-        "user": current_user.user if current_user.user else None,
-        "current_page": result.current_page,
-        "total_pages": result.total_pages,
-        "toys": result.toys,
-        "categories": result.categories,
-        "tags": result.tags,
-    }
-    response: Response = templates.TemplateResponse(
-        request=request,
-        name="toys-catalog.html",
-        context=context
-    )
-
-    for cookie in current_user.cookies:
-        response = set_cookie(response, cookie)
-
-    return response
-
-
-@router.get("/toys/catalog/{toy_id}", response_class=HTMLResponse, name="toy_by_id")
-async def toy_by_id(
-        request: Request,
-        result: ToyByIDResponse = Depends(toy_by_id_dependency),
-        encryptor: FernetEnvironmentsKey = Depends(encryptor_dependency)
-):
-    current_user: GetMeResponse = await get_me_dependency(request=request)
-
-    if result.error is not None:
-        encrypted_error: str = encryptor.encrypt(result.error)
-
-        return RedirectResponse(
-            url=f"/sso/toys/catalog?error={encrypted_error}",
-            status_code=status.HTTP_303_SEE_OTHER
-        )
-
-    context = {
-        "user": current_user.user if current_user.user else None,
-        "toy": result.toy
-    }
-    response: Response = templates.TemplateResponse(request=request, name="toy-page.html", context=context)
-
-    for cookie in current_user.cookies:
-        response = set_cookie(response, cookie)
-
-    return response
