+++ src/toys/constants.py
@@ -1,0 +1,1 @@
+TOYS_PER_PAGE: int = 12
