+++ static/styles/toy-page.css
@@ -1,80 +1,234 @@
+:root {
+    --primary-color: #4CAF50;
+    --hover-color: #45a049;
+    --background: #EDE0D4;
+    --text-color: #1e293b;
+    --border-color: black;
+    --hashtag-bg: #e8f5e9;
+    --card-bg: #F5EDE4;
+    --divider-color: #e0e0e0;
+    --image-bg: #ffffff;
+}
+
+.toy-page-wrapper {
+    padding-top: 30px;
+    background-color: var(--background);
+    min-height: 100vh;
+}
+
 .toy-page-container {
     display: flex;
-    max-width: 1200px;
+    gap: 2rem;
+    max-width: 1400px;
     margin: 0 auto;
-    padding: 20px;
-    gap: 40px;
+    padding: 0 1rem;
 }
 
 .toy-images {
     flex: 1;
+    background: var(--image-bg);
+    border-radius: 8px;
+    padding: 1rem;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.image-gallery {
+    width: 100%;
+}
+
+.image-container {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
 }
 
 .toy-image {
-    width: 100%;
+    max-width: 100%;
     max-height: 500px;
     object-fit: contain;
-    border-radius: 8px;
-    margin-bottom: 20px;
+    border-radius: 4px;
 }
 
 .toy-details {
     flex: 1;
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+
+.toy-card {
+    background: var(--card-bg);
+    border-radius: 8px;
+    padding: 1.5rem;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+}
+
+.seller-card-link {
+    text-decoration: none;
+    color: inherit;
+    display: block;
+}
+
+.seller-card {
+    background: var(--card-bg);
+    border-radius: 8px;
+    padding: 1.5rem;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+    transition: transform 0.2s, box-shadow 0.2s;
+    cursor: pointer;
+}
+
+.seller-card:hover {
+    transform: translateY(-2px);
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.divider {
+    height: 1px;
+    background-color: var(--divider-color);
+    margin: 1rem 0;
+}
+
+.toy-header {
+    margin-bottom: 1rem;
+}
+
+.toy-title {
+    font-size: 2rem;
+    color: var(--text-color);
+    margin: 0 0 0.5rem;
 }
 
 .toy-price {
-    font-size: 2em;
-    color: var(--price-color);
-    margin: 20px 0;
+    font-size: 2.5rem;
+    font-weight: 700;
+    color: var(--primary-color);
+    margin: 0;
 }
 
 .toy-description {
-    margin: 30px 0;
+    font-size: 1.1rem;
+    line-height: 1.6;
+    color: #64748b;
 }
 
-.toy-meta {
-    background: var(--card-bg-color);
-    padding: 20px;
-    border-radius: 8px;
-    margin: 30px 0;
+.toy-description h3, .toy-tags h3, .card-header h3 {
+    font-size: 1.2rem;
+    color: var(--text-color);
+    margin-bottom: 0.5rem;
+}
+
+.toy-tags .tags-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+}
+
+.hashtag {
+    background: var(--hashtag-bg);
+    color: var(--primary-color);
+    padding: 0.5rem 1rem;
+    border-radius: 20px;
+    font-size: 0.9rem;
+    font-weight: 500;
+}
+
+.meta-item {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    padding: 0.5rem 0;
+    color: #64748b;
+}
+
+.meta-item i {
+    color: #78909c;
+    width: 1.5rem;
+    text-align: center;
+}
+
+.seller-info {
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+}
+
+.seller-avatar {
+    width: 56px;
+    height: 56px;
+    border-radius: 50%;
+    object-fit: cover;
+}
+
+.seller-avatar.placeholder {
+    background: #e2e8f0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: #94a3b8;
+}
+
+.seller-details {
+    display: flex;
+    flex-direction: column;
+    gap: 0.25rem;
+}
+
+.seller-name {
+    font-weight: 600;
+    color: var(--text-color);
+}
+
+.seller-id {
+    font-size: 0.875rem;
+    color: #78909c;
+}
+
+.seller-rating {
+    font-size: 0.875rem;
+    color: var(--primary-color);
+}
+
+.seller-rating i {
+    color: #FFC107;
 }
 
 .add-to-cart {
-    padding: 15px 30px;
-    background: var(--btn-bg-color);
-    color: var(--btn-text-color);
+    width: 100%;
+    padding: 1rem;
+    background: var(--primary-color);
+    color: white;
     border: none;
-    border-radius: 8px;
-    font-size: 1.2em;
+    border-radius: 4px;
+    font-size: 1.1rem;
+    font-weight: 600;
     cursor: pointer;
+    transition: background-color 0.3s;
+}
+
+.add-to-cart:hover {
+    background: var(--hover-color);
 }
 
 @media (max-width: 768px) {
     .toy-page-container {
         flex-direction: column;
     }
-}
 
-.toy-seller {
-    margin-bottom: 20px;
-    padding: 15px;
-    background: #f5f5f5;
-    border-radius: 8px;
-}
+    .toy-image {
+        max-height: 350px;
+    }
 
-.seller-info {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
+    .toy-title {
+        font-size: 1.5rem;
+    }
 
-.seller-avatar {
-    width: 50px;
-    height: 50px;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.seller-name {
-    font-weight: 600;
+    .toy-price {
+        font-size: 2rem;
+    }
 }
